Basic pan and zoom image controls.
Prevent the elastic scrolling effect.

Fix the Chrome image zoom performance issue

Chrome is auto-rendering images with bicubic filtering,
that is a bit too slow for interactive pan & zoom.

So, let's drop to nearest neighbor filtering while the
image is in motion, and put it back to high quality
as soon as the image stops moving.

[Finishes #132391221]