Update shared link button colors
Add a hover block so that the text changes to green
when it says "Click to copy"

Change the default colors so that the blue on the
copy link matches the blue for the rest of the button
by default, and then turns dark blue instead of
green on hover. That way, it looks like one big button
by default and you can see which part you are
clicking on when you hover.