Fix for Sidebar resizer preventing interaction in Widgets
Make sure the stacking order of Racebar is after Workspace,
so that children of Racebar stack on top of the Workspace.

This may generally help other z-index issues, since z-index
is being used for drop-downs. Stacking bottom to top allows
dropdowns to automatically stack on top of elements that are
vertically lower than the dropdown parent.